+++ includes/PF_FormField.php
@@ -729,9 +729,11 @@ class PFFormField {
 	/**
 	 * Helper function to get an array of labels from an array of values
 	 * given a mapping property.
+	 * @param \SMW\Store|null $store
+	 * @suppress PhanUndeclaredTypeParameter For Store
 	 */
-	public function setValuesWithMappingProperty() {
-		$store = PFUtils::getSMWStore();
+	public function setValuesWithMappingProperty( $store = null ) {
+		$store ??= PFUtils::getSMWStore();
 		if ( $store == null ) {
 			return;
 		}